requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.7
Flask-WTF==0.14.2
flask-marshmallow==0.9.0
WTForms==2.2.1
web3==4.8.2
py-solc==3.2.0
